requirements.txt
Django>=4.2.0,<5.0.0
Pillow>=10.0.0
gunicorn>=21.2.0
whitenoise>=6.5.0
django-mathfilters
django-widget-tweaks
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0